Fix a test that is erroring with EPIPE
Looks like a new Node version has added an error when piping to a pipe
that doesn't exist, such as stdin for a command that couldn't be found.

Also, sh has been updated with a new error message format, I guess?  One
would think ancient software like that is static, but I guess not!