Update dependency livescript, minimist -> yargs
Just throwing in livescript 1.6.0 made argument handling go wonky for
reasons I still don't fully understand, but moving to yargs is probably
a good idea anyway given minimist hasn't seen updates for years.

It took a bit of experimentation to get this to work when called with no
input files when run as both

$ lsc index.ls

and

$ node index.js

because the argv have to be handled differently.  It annoys me that
that's a thing, but hey I wrote a comment to document.