Clearer output + continue if test-program fails
Turns out we can get much clearer output for this use-case if we
generate the TAP outselves instead of using tape.

It also makes sense to continue testing even if the user-provided test
program fails once.  It failing is still a successful run of the test;
just with a negative result.