Print results as available, maintaining ordering
Previously, instances of the tests of readme examples were run in
parallel, but their results gathered without printing them, and then
presented all at once.  This way it was easy to keep them in the correct
order.

This change prints results as they come in, but only if results for all
tests prior to that one have already been printed.  This gets the best
of both worlds: maintains order, but prints results as soon as they're
available.