Output TAP even for format or test-program errors
Previously the output when the format was wrong (e.g. duplicate in/out
command for some name of test) or when the user-specified test program
failed, the output wouldn't be TAP, and it wouldn't be very useful
either.

Now, output is TAP even when failures like those happen, and the output
shows relevant program outputs and line numbers of where relevant
commands were parsed from, to aid in debugging.