Only require \ to be escaped when it's used as \-
It was a massive headache when writing program commands to escape every
backslash twice (once for the shell, and once for txm), even though all
it's used for at txm level is to escape backslashes.

Previously, all backslashes needed to be escaped.  Now only a a literal
"\-" needs to be escaped (as "\\-").  All other uses of backslash need
no escape.