Add "!test check" command for standalone examples
This is often useful for tests that call `assert` or that otherwise
demonstrate their correctness just by virtue of exiting non-zero.

I made this a separate command instead of just making "!test out"
optional, because it shows intention.  With an optional parameter stuff,
one could easily have a brainfart or make a copypaste mistake, without
it being detected.  Like this feels more reliable.