Remove engines field from package.json
It isn't particularly useful.  I know this needs ES2018 because I'm
using negative lookback in a regex, but I can't be bothered looking up
what Node version that corresponds to right now, and even if I did, I am
not interested in the maintenance load of even testing this with Node
versions that old.