Coverage: Ignore EPIPE error handler
I've tried various ways to try to induce this deliberately in tests, but
none are reliable.

The only way that I've found that would work would require modifying how
the program behaves such that the subprocess is written to on
process.nextTick after it is started, but doing that just for test
coverage doesn't seem worth it.

Instead, I've made comment notes instead, on when the non-determinstic
situation may happen.