Remove yargs dependency, do arg parsing ourselves
Yargs is a kinda big dependency which features we barely use, since our
command line arguments are relatively simple.  So let's do it ourselves.

I tried to keep this about as generic as the previous usage of yargs
was, so adding more flags later is easy, if we need more.

The test change (an exit code change) is because it's conventional at
least on Linux to exit with code 1 on "oops, it died", but code 2 on
"you're using it wrong".