Add `!test exit`, to expect specific exit code
It was previously tedious to do this, because you'd have to temporarily
redefine the testing program to one that modifies its exit code.  For
non-trivial test program scripts, this results in lots of irritating
copy-pasting.

I made it only affect the next test, to encourage being explicit by
repeating it for any test that is supposed to fail.  And because I'd
definitely forget to set it back to 0 and be confused.